refactor to split out code and improve naming
