+++ src/recipes/blocs/RecipeListEvent.tsx
@@ -1,0 +1,7 @@
+// export enum RecipesEvent {
+// fetch = 'FETCH'
+// }
+
+export interface RecipeListEvent {};
+
+export class RecipeListFetchEvent implements RecipeListEvent {};
