+++ src/recipes/blocs/RecipeListStatus.tsx
@@ -1,0 +1,7 @@
+// Note: this could be generic for any list operation
+export enum RecipeListStatus {
+  INITIAL = "initial",
+  PENDING = "pending",
+  SUCCESS = "success",
+  FAILURE = "failure"
+}
