+++ src/recipes/blocs/RecipeStatus.tsx
@@ -1,0 +1,9 @@
+
+export enum RecipeStatus {
+  INITIAL = "initial",
+  PENDING = "pending",
+  CREATED = "created",
+  UPDATED = "updated",
+  DELETED = "deleted",
+  FAILURE = "failure"
+};
